Add and use getSchemeFromUrl util
