+++ lib/rules/function-url-scheme-blacklist/index.js
@@ -2,9 +2,9 @@
 
 const _ = require("lodash")
 const functionArgumentsSearch = require("../../utils/functionArgumentsSearch")
+const getSchemeFromUrl = require("../../utils/getSchemeFromUrl")
 const isStandardSyntaxUrl = require("../../utils/isStandardSyntaxUrl")
 const matchesStringOrRegExp = require("../../utils/matchesStringOrRegExp")
-const parse = require("url").parse
 const report = require("../../utils/report")
 const ruleMessages = require("../../utils/ruleMessages")
 const validateOptions = require("../../utils/validateOptions")
@@ -32,21 +32,9 @@ const rule = function (blacklist) {
           return
         }
         const urlString = _.trim(unspacedUrlString, "'\"")
+        const scheme = getSchemeFromUrl(urlString)
 
-        const url = parse(urlString)
-        if (url.protocol === null) {
-          return
-        }
-
-        const scheme = url.protocol.toLowerCase().slice(0, -1) // strip trailing `:`
-
-        // The URL spec does not require a scheme to be followed by `//`, but checking
-        // for it allows this rule to differentiate <scheme>:<hostname> urls from
-        // <hostname>:<port> urls. `data:` scheme urls are an exception to this rule.
-        const slashIndex = url.protocol.length
-        const expectedSlashes = urlString.slice(slashIndex, slashIndex + 2)
-        const isSchemeLessUrl = expectedSlashes !== "//" && scheme !== "data"
-        if (isSchemeLessUrl) {
+        if (scheme === null) {
           return
         }
